Chatbot/requirements.txt
flask
psycopg2-binary
flask-cors
flask_bcrypt
streamlit
streamlit-chat
langchain
langchain-openai
faiss-cpu
numpy
sentence-transformers
PyPDF2
openai
python-dotenv
gTTS
pinecone
pinecone-client
langchain-community
beautifulsoup4
duckduckgo-search



